+++ projects/mk-kit/translate/index.ts
@@ -1,0 +1,5 @@
+export * from './translate.types';
+export * from './translate.service';
+export * from './translate.pipe';
+export * from './translate.loaders';
+export * from './translate.providers';